+++ app/element/clock.cpp
@@ -7,7 +7,7 @@ bool Clock::reset = false;
 Clock::Clock( QGraphicsItem *parent ) : GraphicElement( 0, 0, 1, 1, parent ) {
   setOutputsOnTop( false );
   setRotatable( false );
-/*  connect(&timer,&QTimer::timeout,this,&Clock::updateClock); */
+  /*  connect(&timer,&QTimer::timeout,this,&Clock::updateClock); */
   setFrequency( 1.0 );
   setHasFrequency( true );
   on = false;
@@ -27,7 +27,6 @@ void Clock::updateClock( ) {
   setOn( on );
 }
 
-
 bool Clock::getOn( ) const {
   return( on );
 }
@@ -42,7 +41,6 @@ void Clock::setOn( bool value ) {
   }
   outputs( ).first( )->setValue( on );
 }
-
 
 void Clock::save( QDataStream &ds ) const {
   GraphicElement::save( ds );
@@ -59,19 +57,19 @@ void Clock::load( QDataStream &ds, QMap< quint64, QNEPort* > &portMap, double ve
 }
 
 float Clock::getFrequency( ) const {
-  return( m_frequency );
+  return( static_cast< float >( m_frequency ) );
 }
 
 void Clock::setFrequency( float freq ) {
-/*  qDebug() << "Clock frequency set to " << freq; */
+  /*  qDebug() << "Clock frequency set to " << freq; */
   if( not qFuzzyIsNull( freq ) ) {
     int auxinterval = 1000 / ( freq * GLOBALCLK );
     if( auxinterval > 0 ) {
       interval = auxinterval;
-      m_frequency = freq;
+      m_frequency = static_cast< double >( freq );
       elapsed = 0;
       Clock::reset = true;
-//      qDebug() << "Freq = " << freq <<  " interval = " << interval;
+      //      qDebug() << "Freq = " << freq <<  " interval = " << interval;
     }
     /*    timer.start( static_cast< int >(1000.0/freq) ); */
   }
@@ -82,7 +80,6 @@ void Clock::resetClock( ) {
   elapsed = 0;
 }
 
-
 QString Clock::genericProperties( ) {
-  return( QString( "%1 Hz" ).arg( getFrequency( ) ) );
+  return( QString( "%1 Hz" ).arg( static_cast< double >( getFrequency( ) ) ) );
 }
