+++ app/elementeditor.cpp
@@ -2,11 +2,11 @@
 #include "editor.h"
 #include "elementeditor.h"
 #include "ui_elementeditor.h"
-#include <cmath>
 #include <QDebug>
 #include <QGraphicsView>
 #include <QKeyEvent>
 #include <QMenu>
+#include <cmath>
 
 ElementEditor::ElementEditor( QWidget *parent ) : QWidget( parent ), ui( new Ui::ElementEditor ) {
   _manyLabels = tr( "<Many labels>" );
@@ -295,7 +295,7 @@ void ElementEditor::setCurrentElements( const QVector< GraphicElement* > &elms )
       if( hasSameFrequency ) {
         ui->doubleSpinBoxFrequency->setMinimum( 0.5 );
         ui->doubleSpinBoxFrequency->setSpecialValueText( QString( ) );
-        ui->doubleSpinBoxFrequency->setValue( firstElement->getFrequency( ) );
+        ui->doubleSpinBoxFrequency->setValue( static_cast< double >( firstElement->getFrequency( ) ) );
       }
       else {
         ui->doubleSpinBoxFrequency->setMinimum( 0.0 );
@@ -423,12 +423,12 @@ bool ElementEditor::eventFilter( QObject *obj, QEvent *event ) {
     if( move_back || move_fwd ) {
       GraphicElement *elm = m_elements.first( );
       QVector< GraphicElement* > elms = scene->getVisibleElements( );
-      std::stable_sort( elms.begin( ), elms.end( ), [ ]( GraphicElement *elm1, GraphicElement *elm2 ) {
-        return( elm1->pos( ).ry( ) < elm2->pos( ).ry( ) );
-      } );
-      std::stable_sort( elms.begin( ), elms.end( ), [ ]( GraphicElement *elm1, GraphicElement *elm2 ) {
-        return( elm1->pos( ).rx( ) < elm2->pos( ).rx( ) );
-      } );
+      std::stable_sort( elms.begin( ), elms.end( ), [] ( GraphicElement * elm1, GraphicElement * elm2 ) {
+                          return( elm1->pos( ).ry( ) < elm2->pos( ).ry( ) );
+                        } );
+      std::stable_sort( elms.begin( ), elms.end( ), [] ( GraphicElement * elm1, GraphicElement * elm2 ) {
+                          return( elm1->pos( ).rx( ) < elm2->pos( ).rx( ) );
+                        } );
 
       apply( );
       int elmPos = elms.indexOf( elm );
